Add files via upload
5.ReportCard (Required)
Project Rubric
Students will design a custom Java class to represent a school report card from an educational system of their choice. This project is not a functional app. Rather, it is practice designing custom classes to represent real world objects and behavior. 
Upon completing this project, students will be able to:
•	Design a custom java class. 
•	Write the class code, including local variables, getter and setter methods, and methods that make sense for the object in question. 
•	Comment their code.